removed duplicate useEffect in chart.js
